Setup Hungarian Y and Z position in English too
Change the LANG key so that it not only sends a Super+SPACE key combination but it also toggles an additional layer, which exchanges Y and Z relative to the standard QWERTY layout.
If the OS input language and the keyboard is correctly synchronized, the keyboard layout is QWERTZ in both English and Hungarian.