Set bot bottom vertical 1,5 key to AltGr, left bottom row second key ESC
I have not found tappable Shift and AltGr usable on a frequently used letter key. As it turns out, I frequently press the next key before I release the previous one. If the tappable key is a frequently used letter, then pressing that letter quickly followed by another letter will be recognized as if only the second were pressed together with the modifier assigned to the first key.